+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.5.0] - 2022-12-28
+
+### Added
+
+- Add button to display history of each daily sale origin (cargo, shop, etc)
+
+### Changed
+
+- Last sales by item show now 1,000 entries instead of 50
+
 ## [2.4.0] - 2022-09-14
 
 ### Added